add color code superclass
